Don't mutate activeStateResolveContent before states are destroyed
This bug was introduced in the change where state resets went away – when states are destroyed, their resolve data on the global activeStateResolveContent object would get wiped.

Now that resets have been converted into destroys+creates, this one line that tosses the data from the resolve functions onto activeStateResolveContent is obviously in the wrong place... it happens right before the states get destroyed.

If any of the states that were destroyed are also being created, the data returned by the resolve functions for the created states would get ignored.